agregado alerta de confirmacion de sweet alert

+++ src/components/Home.js
@@ -2,8 +2,10 @@ import React, { useState, useEffect } from 'react';
 
 import firebaseApp from '../credenciales';
 import { getAuth, signOut } from 'firebase/auth';
-
 import { getFirestore, collection, addDoc, getDocs, doc, deleteDoc, getDoc, setDoc } from 'firebase/firestore';
+
+import Swal from 'sweetalert2/dist/sweetalert2.js';
+import '@sweetalert2/theme-bootstrap-4/bootstrap-4.css'
 
 const auth = getAuth(firebaseApp);
 const db = getFirestore(firebaseApp);
@@ -31,19 +33,35 @@ export const Home = ({ correoUsuario }) => {
 	const guardarDatos = async (e) => {
 		e.preventDefault();
 
-		if (subId === '') {
-			try {
-				await addDoc(collection(db, 'usuarios'), {
-					...user
-				})
-			} catch (error) {
-				console.log(error);
+		Swal.fire({
+			title: 'Do you want to save the changes?',
+			showDenyButton: true,
+			showCancelButton: true,
+			confirmButtonText: 'Save',
+			denyButtonText: `Don't save`,
+		}).then((result) => {
+			if (result.isConfirmed) {
+				if (subId === '') {
+					try {
+						addDoc(collection(db, 'usuarios'), {
+							...user
+						})
+					} catch (error) {
+						console.log(error);
+					}
+				} else {
+					setDoc(doc(db, 'usuarios', subId), {
+						...user
+					});
+				}
+				Swal.fire('Saved!',
+					'',
+					'success'
+				)
+			} else if (result.isDenied) {
+				Swal.fire('Changes are not saved', '', 'info')
 			}
-		} else {
-			await setDoc(doc(db, 'usuarios', subId), {
-				...user
-			});
-		}
+		})
 		setUser({ ...valorInicial });
 		setSubId('');
 	}
@@ -67,7 +85,26 @@ export const Home = ({ correoUsuario }) => {
 
 	// Funcion para eliminar un usuario
 	const deleteUser = async (id) => {
-		await deleteDoc(doc(db, 'usuarios', id));
+
+		Swal.fire({
+			title: 'Estas seguro?',
+			text: "No podrás revertir este cambio",
+			icon: 'warning',
+			showCancelButton: true,
+			confirmButtonColor: '#3085d6',
+			cancelButtonColor: '#d33',
+			confirmButtonText: 'Si, eliminar!',
+			cancelButtonText: 'Cancelar'
+		}).then((result) => {
+			if (result.isConfirmed) {
+				deleteDoc(doc(db, 'usuarios', id));
+				Swal.fire(
+					'Eliminado!',
+					'El cliente fue eliminado.',
+					'success'
+				)
+			}
+		})
 	}
 
 	const getOne = async (id) => {
@@ -84,21 +121,40 @@ export const Home = ({ correoUsuario }) => {
 		if (subId !== '') {
 			getOne(subId)
 		}
-	}, [subId])
+	}, [subId]);
+
+	// Alertas de verificación Sweet Alert
+	const alert = () => {
+		Swal.fire({
+			title: '¿Estas seguro de salir?',
+			icon: 'warning',
+			showCancelButton: true,
+			confirmButtonColor: '#3085d6',
+			cancelButtonColor: '#d33',
+			confirmButtonText: 'Sí, cerrar sesión',
+			cancelButtonText: 'Cancelar'
+		}).then((result) => {
+			if (result.value) {
+				Swal.fire(
+					signOut(auth)
+				)
+			}
+		})
+	}
 
 	return (
 		<div className='container'>
 			<p>Bienvenido, <strong>{correoUsuario}</strong> Haz iniciado sesión.</p>
 
-			<button className='btn btn-primary' onClick={() => signOut(auth)}>
+			<button className='btn btn-primary btn-sm' onClick={() => alert()}>
 				Cerrar Sesión.
 			</button>
 
 			<hr />
 
-			<div className='row'>
+			<div className='row mt-5'>
 				{/* Seccion Formulario */}
-				<div className="col-md-4">
+				<div className="col-md-4 mt-3">
 					<h3 className='text-center mb-3'>Ingresar usuario.</h3>
 					<form onSubmit={guardarDatos}>
 						<div className='card card-body'>
@@ -118,7 +174,7 @@ export const Home = ({ correoUsuario }) => {
 				</div>
 
 				{/* Seccion lista de usuarios */}
-				<div className="col-md-8">
+				<div className="col-md-8 mt-3">
 					<h2 className='text-center mb-3'>Lista de Usuarios</h2>
 
 					<div className="container card">
@@ -126,6 +182,7 @@ export const Home = ({ correoUsuario }) => {
 							{
 								lista.map(list => (
 									<div key={list.id}>
+										<hr />
 										<p>Nombre: {list.nombre}</p>
 										<p>Edad: {list.edad}</p>
 										<p>Profesión: {list.profesion}</p>
@@ -136,7 +193,6 @@ export const Home = ({ correoUsuario }) => {
 										<button className='btn btn-success ms-4' onClick={() => setSubId(list.id)}>
 											Actualizar
 										</button>
-										<hr />
 									</div>
 								))
 							}
